Only ask for root permission on necessary command
Before it was checked for all commands (so even the commands not needing
root permissions as all getter) if the binary is executed as root, now
it is only done for the set commands (which manipulate the efivars).
Also removed the file attribute restore on `Drop` for efivars since it
would prevent to run without root even for get commands.
This is not considered a security issue as you anyways need root to
manipulate the files and as root you can also remove the immutable bit.