Fix initial check for whether OTA downloads allowed (#35)
When the orb first booted, the downloads would throttle for the first
DEFAULT_DURATION_TO_ALLOW_DOWNLOADS duration, since we would fill the
queue with an `Instant::now()`. The behavior is changed by subtracting
the default duration from the now to ensure the elapsed time is correct
on first check.