ci: autodetect cargo deb crates (#101)
Lets us stop having to manually write out the list of binaries. Also
will allow us to reuse this workflow in orb-internal, since it is no
longer specific to this repo.

Also seeking opinions: Would it be better for me to keep the json
parsing in jq, or do you want me to port it to native python?