can: fix ack number with a high pid
ack number uses the pid in the 16 most significant bits
(and a counter in the lower 16 bits),
this allows us to filter out acks that are not
for the current process.
But, the pid can be encoded over more than 16 bits,
so this fix allows the usage of the 16 lsb of the pid
in the ack number, the msb of the pid are lost.