feat: migrate fleet_cmdr client to hybrid orchestrator pattern (#567)
The purpose of this PR is to modernize the "fleet cmdr" client on the
orb by aligning it with the current Jobs protocol design for our
backend. This involves adding support for job cancellation, processing
of multiple jobs in parallel, and also ensuring that we request another
job as soon as one completes.

It also converts the naming to "orb-jobs-agent" because we ended up
using the Fleet Cmdr name for an internal application.