feat(hil): expose nfs (#573)
This PR does several things:
* adds the necessary packages to the `tegra-bash` FHSEnv to enable the
use of the boot-from-ram bootloader0 scripts.
* Configures a known name for the network interface of the orb's RCM
ethernet, based on VID and PID, and names it `orbeth0`. This allows us
to replicate our configuration across many orbs and NUCs.
* Configures a networkmanager connection for `orbeth0`, in particular
setting it as a "shared" connection, which will cause the orb to run a
DHCP server to assign ip addresses to the orb - necessary for NFS.
* Configures the NUC to run a NFS server. Anything under `/srv` will be
exposed to the orb over NFS, and only on the local ipv4 10.42.0.0/24
subnet.
* Configures port forwarding for the NFS server.