chore(mcu-util): send one can-fd message (#577)
to trigger broadcast of all the messages over can-fd as this will add a
subscriber.
Regardless of the preferred communication protocol (isotp or canfd).

since worldcoin-mcu-telemetry starts only when internet connection is
alive, we need orb-mcu-util to trigger broadcast of all the messages so
that battery can be known (relying on messages sent every second)