refactor(update-agent-loader): pass configuration in explicitly (#650)
Following changes were made:
* add `orb-` prefix, as conventional in our workspace
* set version to 0.0.0. as conventional in our workspace
* Added a `Config` struct which has a `from_env` function, call this in
`main`, and pass this into the rest of the code to be explicit about
configuration of the program (and make it more possible to test the
code)
* Removed `build.rs` as it both doesn't make sense to check target os
and error (the code will already error if its platform specific) and was
non-hermetic (randomly generates build-time compiled constants from a
non-seeded RNG)
* As per @alekseifedotov's request, I kept the constraint that if a
public key is not provided at build time, a randomly generated private
key will be used. the difference now is that this random generation
happens at run time, not build time. And with the dependency injection,
it is now possible to see that random pubkey in tests.