fix(backend-status,location): fix tempfile bug in tests (#714)
Fixed several bugs:
* tempfile that was created in a local scope is dropped, meaning the
file doesn't exist (causing errors on my machine)
* the path to the tempfile was incorrectly prefixed with /tmp
* the use of a tempfile + '.listen()` at all is unecessary, the
`dbus-launch` crate already handles this.
* Indicating the bus type (session in this case) is the ideal way to
handle that
* for some reason, the tests "fail successfully" when dbus-launch is not
detected on the commandline. Not having it on the commandline should be
considered a bug, and also there are better ways to exclude tests in
cases such as when running inside docker (not that this is relevant
here)