feat(s3-helpers): implement multipart upload (#713)
Vibe coded with codex + gpt5. I have reviewed the code.

It did actually well, specifically because I told it to mimic the same
styles used in my manually-implemented multipart_download from prior
PRs. I think the approach of "manually implement it first, get it good,
then have the AI clone the pattern" is a good way to use things. This is
a different approach to how I usually see it used, which is "please go
do this feature" (which typically results in pretty poor code).

Was pretty impressed that it oneshot it. I only did 1 minor nitpick
tweak to switch the page_number to NonZeroU16. It probably helps that
doing an s3 upload is an *extremely* common task.