feat(orb-jobs-agent): add service command (#791)
# Add a `service` command to `orb-jobs-agent`

## Summary 
Adds a new `service` command to orb-jobs-agent that allows Fleet
Commander to control systemd services on the Orb. This enables remote
management of services (start, stop, restart, status).

## Supported Actions

- `start`
- `stop`
- `restart`
- `status`

```sh
service stop worldcoin-core.service
```

## Notes
There was a concern about an injection in the command that would look
something like this

```sh
service stop worldcoin.core-serivce;shutdown <----
```
This is not possible with the current implementation of the `Shell`
trait for `Host`. It does not span an actual shell, that would be able
to parse an input like that, but rather executes command directly, so in
the example above `worldcoin.core-serivce;shutdown` would be treated as
an arg to `systemctl` and it will fail with no such service.