Revert "chore(connd): change connectivity check to http (#794)" (#795)
This reverts commit 015295e0e8ae2e670abee98dd7a73d85bc2d36cf.

Using `http` will cause NM to identify connection status as `portal` as
opposed to `full` since we use Cloudflare to do a domain redirect on the
uri check. Using `https` will return a `204` instead of a `301` and then
a `204`.

Only problem right now is using `https` will cause the connectivity
check to fail if we have any NTP issues.