feat(orb-hil): fix ota command (#815)
# Fixes a bug in OTA command and some other improvements

# Bug Fix
Fixes the bug of OTA command failing due to reading the `journalctl`
logs of previously failed runs. Now reads the time on the orb and uses
`--since` flag to make sure the logs are relevant to current update.

# Changes

- Refactored OTA command, which became too long to be a standalone file
into multiple modules
- Now we do not rely on output from the serial to login after reboot (
this can be flaky and some line are lost). We have timeout and try to
SSH even if we don't get the login prompt from serial
- Switched to using `orb-mcu-util reboot orb;shutdown` instead of reboot
( to save retry counts during testing/failed OTAs)
- Periodically check `update-agent status,` not parse logs for failure

Tested on the HIL and diamond orb multiple times, not tested on pearl
yet

---------

Co-authored-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>
Co-authored-by: vmenge <victor.ferreira.menge@toolsforhumanity.com>
Co-authored-by: AlexKaravaev <30314738+AlexKaravaev@users.noreply.github.com>