feat(secure-storage): implement TA user auth (#862)
It is nicer to use the effective user id of the CA instead of an opaque
UUID in the TA, and its also more straightforward to work with.

Since REE only sees provides the TA a UUID [derived from the CA's
effective user
id](https://elixir.bootlin.com/linux/v5.15.148/source/drivers/tee/tee_core.c#L232),
we pass in the euid as a parameter to the TA when the session is crated,
map that to a UUID, and then validate that it matches the expected uuid.

Note that the euid is still ultimately controlled by the linux kernel,
so the TA must factor that into its trust model.