feat(connd): add code for subprocess worker (#873)
The code is complete although perhaps hastily written. Its possible
there are some spots that are missing implementation or are bugged.

We also need to:
* figure out why tests fail on my machine
* figure out if we need cargo escargot for execing the subprocess while
inside a test case
* support runtime configuration for determining if we are to use the
in-memory stubbed backend for orb-secure-storage-ca or if instead we use
the actual `OpteeBackend`. Right now I just default to in-memory for
ease of testing.
* imo we should not be using multthreaded tokio runtime in child worker
process
* Do we need to clean up the EntryPoint code?

---------

Co-authored-by: vmenge <victor.ferreira.menge@toolsforhumanity.com>