refactor(orb-backend-status): improve reliability, testing capabilities and logic (#876)
# Refactor `orb-backend-status`

This PR refactors `orb-backend-status` into a more modular and testable
implementation, fixes reliability issues and adds some new changes.
Partially tested on the orb, need the test recent changes again.

I understand this is a big PR to review a lot of stuff was changes, so I
had a 1:1 review call with @vmenge and am okay with making another one
if someone wants to go through this

# Main changes

1. Serve Dbus as soon as possible, so that other crates can access it
2. Do not fail anywhere (e.g. token missing, no connectivity ) 
3. Add logic for urgent notifications that are easy to add
4. Add integration tests

# Service core loop

1. Regular updates are polled + sent every 30 seconds
2. If something is urgent ( currently only reboot notification from
update-agent, or SSID change from connd) -> send immediately
3. Sending is gated by connectivity and token. If no internet or no
token, the service will not try to send, and wait for conditions to be
met
4. If sending fails (e.g. backend returns 500) it will retry with a
backoff

---------

Co-authored-by: vmenge <victor.ferreira.menge@toolsforhumanity.com>