feat(secure-storage): introduce storage domains (#888)
Best practices dictate that each "use case" have its own TA, since each
TA's storage is isolated from all other TAs. As a first step towards
implementing this, this PR introduces the concept of a "storage domain",
such that in addition to being scoped by EUID, the storages are also
scoped by domain.

Right now, this is just a client-side abstraction. No changes to the TAs
are necesssary yet because there is only 1 domain. When we wish to add
additional domains (for any other secrets) we will need to library-ify
the TA and make multiple.